requirements.txt
streamlit==1.41.1
cvzone==1.6.1
opencv-python==4.10.0.84
numpy==1.23.5
pillow==9.5.0
ultralytics==8.3.40
torchvision==0.20.1